+++ .gitignore
@@ -5,6 +5,5 @@ database.ini
 backup/
 __pycache__/
 *.pyc
-**/*/karma.py
 **/*/test2.py
 admin_scripts
